Move authActions to the actions folder
We probably don't need a lot of level of nesting
